+++ src/Framework/tests/Routing/Commands/Mocks/MiddlewareB.php
@@ -12,9 +12,21 @@ declare(strict_types=1);
 
 namespace Aphiria\Framework\Tests\Routing\Commands\Mocks;
 
+use Aphiria\Middleware\IMiddleware;
+use Aphiria\Net\Http\IRequest;
+use Aphiria\Net\Http\IRequestHandler;
+use Aphiria\Net\Http\IResponse;
+
 /**
  * Mocks a middleware class
  */
-class MiddlewareB
+class MiddlewareB implements IMiddleware
 {
+    /**
+     * @inheritdoc
+     */
+    public function handle(IRequest $request, IRequestHandler $next): IResponse
+    {
+        return $next->handle($request);
+    }
 }